Refactoring redis implementation for dependecy isolation
