+++ ak-py/pyproject.toml
@@ -18,7 +18,6 @@ dependencies = [
     "pydantic-settings>=2.10.1",
     "pyyaml>=6.0.2",
     "singleton-type>=0.0.5",
-    "redis>=6.4.0",
 ]
 
 [project.optional-dependencies]
@@ -45,6 +44,9 @@ cli = [
 ]
 aws = [
 
+]
+redis = [
+    "redis>=7.1.0",
 ]
 openai = [
     "openai-agents>=0.2.3",
