+++ ak-py/src/agentkernel/core/__init__.py
@@ -16,4 +16,3 @@ from .config import AKConfig as Config
 from .module import Module
 from .runtime import GlobalRuntime, Runtime
 from .service import AgentService
-from .sessions.redis import RedisDriver, RedisSessionSerde
