+++ ak-py/src/agentkernel/core/builder.py
@@ -2,9 +2,7 @@ import logging
 from enum import StrEnum
 
 from .config import AKConfig
-from .sessions import SessionStore
-from .sessions.in_memory import InMemorySessionStore
-from .sessions.redis import RedisDriver, RedisSessionStore
+from .session import SessionStore
 
 
 class Builder:
@@ -79,6 +77,8 @@ class SessionStoreBuilder(Builder):
 
         Builder._log.info(f"Building {session_store_type} session store")
         if session_store_type == SessionStoreType.REDIS:
-            return RedisSessionStore(RedisDriver())
+            from .session.redis import RedisSessionStore
+            return RedisSessionStore()
         else:
+            from .session.in_memory import InMemorySessionStore
             return InMemorySessionStore()
