+++ ak-py/src/agentkernel/core/runtime.py
@@ -9,7 +9,7 @@ from singleton_type import Singleton
 
 from .base import Agent, Session
 from .builder import SessionStoreBuilder
-from .sessions import SessionStore
+from .session import SessionStore
 
 
 class Runtime:
