+++ ak-py/src/agentkernel/core/session/__init__.py
@@ -12,5 +12,3 @@ except importlib.metadata.PackageNotFoundError:
     __version__ = "0.1.0"
 
 from .base import SessionStore
-from .in_memory import InMemorySessionStore
-from .redis import RedisSessionStore
