+++ ak-py/src/agentkernel/core/session/base.py
