+++ examples/aws-serverless/crewai/pyproject.toml
@@ -5,7 +5,7 @@ description = "Demo of CrewAI based agents running in Agent Kernel on AWS Lambda
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "agentkernel[crewai,aws]>=0.2.5"
+    "agentkernel[crewai,redis]>=0.2.5"
 ]
 
 [dependency-groups]
