+++ examples/aws-serverless/langgraph/deploy/deploy.sh
@@ -10,7 +10,7 @@ create_deployment_package() {
       uv pip install -r requirements.txt --target=dist/data
     else
       uv pip install -r requirements.txt --target=dist/data  --find-links ../../../ak-py/dist
-      uv pip install --force-reinstall --target=dist/data --find-links ../../../ak-py/dist agentkernel[langgraph,aws] || true
+      uv pip install --force-reinstall --target=dist/data --find-links ../../../ak-py/dist agentkernel[langgraph,redis] || true
     fi
     cp -r lambda.py custom_agent.py config.yaml dist/data
     popd || exit 1
