+++ examples/aws-serverless/openai/config.yaml
@@ -3,4 +3,4 @@ session:
   type: redis
   redis:
     prefix: "ak:openai:example:"
-    url: "redis://dev-examples-redis.qaandw.0001.apse2.cache.amazonaws.com:6379"
+    url: "redis://ak-openai-serverless-dev-examples-redis.qaandw.0001.apse2.cache.amazonaws.com:6379"
